Variables, types, arrays and loops
Introduces a fixed-size array of grades, which have the implicit double
type. We can simply use `var` when giving a value to a variable, which
automatically sets the data type for us. In the same way, we can use
`new[]` as opposed to `new double[]` as we set the value directly.

Uses a `foreach` loop to add-up all the numbers in our array and return
a result in the console, as a double.